@@ -5,8 +5,8 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.2 <4.0.0'
-  flutter: ">=3.10.0"
+  sdk: ">=3.7.0 <4.0.0"
+  flutter: ">=3.29.0"
 
 dependencies:
   flutter:
@@ -15,7 +15,7 @@ dependencies:
     path: ../
   intl: ^0.20.2
   flutter_svg: ^2.0.9
-  go_router: ^14.6.2
+  go_router: ^17.0.0
   url_launcher: ^6.3.1
 
 dev_dependencies:
