@@ -20,7 +20,10 @@ class BubbleSizeGuide {
 
 /// Widget that renders a chart legend
 class LegendWidget extends StatefulWidget {
-  final List<LegendItem> items;
+  final String? yTitle;
+  final List<LegendItem> itemsY;
+  final String? y2Title;
+  final List<LegendItem> itemsY2;
   final LegendConfig config;
   final ChartTheme theme;
   final BubbleSizeGuide? bubbleGuide;
@@ -34,7 +37,10 @@ class LegendWidget extends StatefulWidget {
 
   const LegendWidget({
     super.key,
-    required this.items,
+    required this.yTitle,
+    required this.itemsY,
+    required this.y2Title,
+    required this.itemsY2,
     required this.config,
     required this.theme,
     this.bubbleGuide,
@@ -112,13 +118,17 @@ class _LegendWidgetState extends State<LegendWidget> {
 
   @override
   Widget build(BuildContext context) {
-    if (widget.items.isEmpty && widget.bubbleGuide == null) {
+    if (widget.itemsY.isEmpty &&
+        widget.itemsY2.isEmpty &&
+        widget.bubbleGuide == null) {
       return const SizedBox.shrink();
     }
 
     // Create default text style using theme colors
-    final defaultTextStyle =
-        TextStyle(fontSize: 12, color: widget.theme.axisColor);
+    final defaultTextStyle = TextStyle(
+      fontSize: 12,
+      color: widget.theme.axisColor,
+    );
 
     // Merge with custom text style, preserving theme color if no color is specified
     final effectiveTextStyle = widget.config.textStyle != null
@@ -155,7 +165,10 @@ class _LegendWidgetState extends State<LegendWidget> {
 
   /// Check if data series items should be stacked vertically next to bubble guide
   bool _shouldStackColorItems(TextStyle textStyle) {
-    if (widget.bubbleGuide == null || widget.items.isEmpty) return false;
+    if (widget.bubbleGuide == null ||
+        (widget.itemsY.isEmpty && widget.itemsY2.isEmpty)) {
+      return false;
+    }
 
     // Calculate bubble guide height with validated size
     final rawMaxBubbleRadius = widget.bubbleGuide!.sizeScale.scale(
@@ -182,8 +195,9 @@ class _LegendWidgetState extends State<LegendWidget> {
     final itemHeight = widget.config.symbolSize > baseFontSize
         ? widget.config.symbolSize
         : baseFontSize;
-    final totalItemsHeight = (itemHeight * widget.items.length) +
-        (widget.config.itemSpacing * (widget.items.length - 1));
+    final itemsLength = widget.itemsY.length + widget.itemsY2.length;
+    final totalItemsHeight = (itemHeight * itemsLength) +
+        (widget.config.itemSpacing * (itemsLength - 1));
 
     return totalItemsHeight <= bubbleGuideHeight;
   }
@@ -196,33 +210,65 @@ class _LegendWidgetState extends State<LegendWidget> {
       widgets.add(_buildHorizontalBubbleGuide(textStyle));
     }
 
-    // Add color legend items - stack if there's a bubble size guide and
-    //they fit, otherwise wrap
-    if (widget.items.isNotEmpty) {
-      final shouldStack = _shouldStackColorItems(textStyle);
+    final shouldStack = _shouldStackColorItems(textStyle);
 
-      widgets.add(
-        shouldStack
-            ? Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: widget.items
-                    .map((item) => Padding(
+    List<Widget> makeLegendItems(
+      List<LegendItem> items,
+      String? title,
+      TextStyle textStyle,
+      bool shouldStack,
+    ) {
+      final legendWidgets = <Widget>[];
+
+      if (title != null && items.isNotEmpty && widget.config.showTitles) {
+        legendWidgets.add(
+          Padding(
+            padding: const EdgeInsets.only(bottom: 8.0),
+            child: Text(
+              title,
+              style: textStyle.copyWith(fontWeight: FontWeight.bold),
+            ),
+          ),
+        );
+      }
+
+      // Add color legend items - stack if there's a bubble size guide and
+      //they fit, otherwise wrap
+      if (items.isNotEmpty) {
+        legendWidgets.add(
+          shouldStack
+              ? Column(
+                  mainAxisSize: MainAxisSize.min,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: items
+                      .map(
+                        (item) => Padding(
                           padding: EdgeInsets.only(
-                              bottom: widget.config.itemSpacing / 2),
+                            bottom: widget.config.itemSpacing / 2,
+                          ),
                           child: _buildLegendItem(item, textStyle),
-                        ))
-                    .toList(),
-              )
-            : Wrap(
-                spacing: widget.config.itemSpacing,
-                runSpacing: widget.config.itemSpacing / 2,
-                children: widget.items
-                    .map((item) => _buildLegendItem(item, textStyle))
-                    .toList(),
-              ),
-      );
+                        ),
+                      )
+                      .toList(),
+                )
+              : Wrap(
+                  spacing: widget.config.itemSpacing,
+                  runSpacing: widget.config.itemSpacing / 2,
+                  children: items
+                      .map((item) => _buildLegendItem(item, textStyle))
+                      .toList(),
+                ),
+        );
+      }
+      return legendWidgets;
     }
+
+    widgets.addAll(
+      makeLegendItems(widget.itemsY, widget.yTitle, textStyle, shouldStack),
+    );
+    widgets.addAll(
+      makeLegendItems(widget.itemsY2, widget.y2Title, textStyle, shouldStack),
+    );
 
     return Wrap(
       spacing: widget.config.itemSpacing * 2,
@@ -239,17 +285,56 @@ class _LegendWidgetState extends State<LegendWidget> {
     // Add bubble size guide first (top) if present
     if (widget.bubbleGuide != null) {
       widgets.add(_buildVerticalBubbleGuide(textStyle));
-      if (widget.items.isNotEmpty) {
+      if (widget.itemsY.isNotEmpty || widget.itemsY2.isNotEmpty) {
         widgets.add(SizedBox(height: widget.config.itemSpacing));
       }
     }
 
+    if (widget.yTitle != null &&
+        widget.itemsY.isNotEmpty &&
+        widget.config.showTitles) {
+      widgets.add(
+        Padding(
+          padding: const EdgeInsets.only(bottom: 8.0),
+          child: Text(
+            widget.yTitle!,
+            style: textStyle.copyWith(fontWeight: FontWeight.bold),
+          ),
+        ),
+      );
+    }
+
     // Add color legend items
     widgets.addAll(
-      widget.items.map((item) => Padding(
-            padding: EdgeInsets.only(bottom: widget.config.itemSpacing),
-            child: _buildLegendItem(item, textStyle),
-          )),
+      widget.itemsY.map(
+        (item) => Padding(
+          padding: EdgeInsets.only(bottom: widget.config.itemSpacing),
+          child: _buildLegendItem(item, textStyle),
+        ),
+      ),
+    );
+
+    if (widget.y2Title != null &&
+        widget.itemsY2.isNotEmpty &&
+        widget.config.showTitles) {
+      widgets.add(
+        Padding(
+          padding: const EdgeInsets.only(bottom: 8.0),
+          child: Text(
+            widget.y2Title!,
+            style: textStyle.copyWith(fontWeight: FontWeight.bold),
+          ),
+        ),
+      );
+    }
+
+    widgets.addAll(
+      widget.itemsY2.map(
+        (item) => Padding(
+          padding: EdgeInsets.only(bottom: widget.config.itemSpacing),
+          child: _buildLegendItem(item, textStyle),
+        ),
+      ),
     );
 
     return Column(
@@ -398,13 +483,25 @@ class _LegendWidgetState extends State<LegendWidget> {
           children: [
             // Bubble guide always shows three bubbles: min, median, and max from the domain
             _buildBubbleItem(
-                displayMinSize, formatValue(minValue), guide.color, textStyle),
+              displayMinSize,
+              formatValue(minValue),
+              guide.color,
+              textStyle,
+            ),
             SizedBox(width: widget.config.itemSpacing),
             _buildBubbleItem(
-                displayMidSize, formatValue(midValue), guide.color, textStyle),
+              displayMidSize,
+              formatValue(midValue),
+              guide.color,
+              textStyle,
+            ),
             SizedBox(width: widget.config.itemSpacing),
             _buildBubbleItem(
-                displayMaxSize, formatValue(maxValue), guide.color, textStyle),
+              displayMaxSize,
+              formatValue(maxValue),
+              guide.color,
+              textStyle,
+            ),
           ],
         ),
       ],
@@ -468,21 +565,40 @@ class _LegendWidgetState extends State<LegendWidget> {
         ),
         // Bubble guide always shows three bubbles: min, median, and max from the domain
         SizedBox(height: LegendWidget._bubbleTitleSpacing),
-        _buildBubbleItemWithAlignment(displayMinSize, formatValue(minValue),
-            guide.color, maxDiameter, textStyle),
+        _buildBubbleItemWithAlignment(
+          displayMinSize,
+          formatValue(minValue),
+          guide.color,
+          maxDiameter,
+          textStyle,
+        ),
         SizedBox(height: widget.config.itemSpacing / 4),
-        _buildBubbleItemWithAlignment(displayMidSize, formatValue(midValue),
-            guide.color, maxDiameter, textStyle),
+        _buildBubbleItemWithAlignment(
+          displayMidSize,
+          formatValue(midValue),
+          guide.color,
+          maxDiameter,
+          textStyle,
+        ),
         SizedBox(height: widget.config.itemSpacing / 4),
-        _buildBubbleItemWithAlignment(displayMaxSize, formatValue(maxValue),
-            guide.color, maxDiameter, textStyle),
+        _buildBubbleItemWithAlignment(
+          displayMaxSize,
+          formatValue(maxValue),
+          guide.color,
+          maxDiameter,
+          textStyle,
+        ),
       ],
     );
   }
 
   /// Build a single bubble item (for horizontal layout)
   Widget _buildBubbleItem(
-      double size, String value, Color color, TextStyle textStyle) {
+    double size,
+    String value,
+    Color color,
+    TextStyle textStyle,
+  ) {
     final diameter = size * 2;
     // fontSize is guaranteed non-null from effectiveTextStyle in build()
     final baseFontSize = textStyle.fontSize!;
@@ -501,19 +617,19 @@ class _LegendWidgetState extends State<LegendWidget> {
           ),
         ),
         SizedBox(height: LegendWidget._bubbleLabelSpacing),
-        Text(
-          value,
-          style: textStyle.copyWith(
-            fontSize: labelFontSize,
-          ),
-        ),
+        Text(value, style: textStyle.copyWith(fontSize: labelFontSize)),
       ],
     );
   }
 
   /// Build a single bubble item with alignment (for vertical layout)
-  Widget _buildBubbleItemWithAlignment(double size, String value, Color color,
-      double maxDiameter, TextStyle textStyle) {
+  Widget _buildBubbleItemWithAlignment(
+    double size,
+    String value,
+    Color color,
+    double maxDiameter,
+    TextStyle textStyle,
+  ) {
     final diameter = size * 2;
     // fontSize is guaranteed non-null from effectiveTextStyle in build()
     final baseFontSize = textStyle.fontSize!;
@@ -537,12 +653,7 @@ class _LegendWidgetState extends State<LegendWidget> {
           ),
         ),
         const SizedBox(width: 6),
-        Text(
-          value,
-          style: textStyle.copyWith(
-            fontSize: labelFontSize,
-          ),
-        ),
+        Text(value, style: textStyle.copyWith(fontSize: labelFontSize)),
       ],
     );
   }
@@ -551,13 +662,17 @@ class _LegendWidgetState extends State<LegendWidget> {
 /// Utility class to generate legend items from chart data and configuration
 class LegendGenerator {
   /// Generate legend items from chart data and color column
-  static List<LegendItem> generateFromData({
+  static (List<LegendItem>, List<LegendItem>) generateFromData({
     required List<Map<String, dynamic>> data,
     required String? colorColumn,
+    required String? yColumn,
+    required String? y2Column,
     required List<Color> colorPalette,
     required List<Geometry> geometries,
   }) {
-    if (colorColumn == null || data.isEmpty || colorPalette.isEmpty) return [];
+    if (colorColumn == null || data.isEmpty || colorPalette.isEmpty) {
+      return ([], []);
+    }
 
     // Extract unique categories from the color column
     final categories = data
@@ -567,28 +682,68 @@ class LegendGenerator {
         .toSet()
         .toList();
 
-    if (categories.isEmpty) return [];
+    if (categories.isEmpty) return ([], []);
 
     // Determine the appropriate symbol based on geometries
     final symbol = _determineSymbolFromGeometries(geometries);
 
-    // Create legend items
-    return categories.asMap().entries.map((entry) {
-      final index = entry.key;
-      final category = entry.value;
-      final color = colorPalette[index % colorPalette.length];
+    // Separate geometries by Y-axis
+    final primaryGeometries =
+        geometries.where((g) => g.yAxis == YAxis.primary).toList();
+    final secondaryGeometries =
+        geometries.where((g) => g.yAxis == YAxis.secondary).toList();
 
-      return LegendItem(
-        label: category,
-        color: color,
-        symbol: symbol,
+    // Generate legend items for primary Y-axis (yColumn)
+    final itemsY = <LegendItem>[];
+    if (yColumn != null && primaryGeometries.isNotEmpty) {
+      // Filter categories that have data in the primary Y column
+      final primaryCategories = categories.where((category) {
+        return data.any(
+          (row) =>
+              row[colorColumn]?.toString() == category && row[yColumn] != null,
+        );
+      }).toList();
+
+      // Use global index of original categories to maintain consistent colors
+      itemsY.addAll(
+        primaryCategories.map((category) {
+          final globalIndex = categories.indexOf(category);
+          final color = colorPalette[globalIndex % colorPalette.length];
+
+          return LegendItem(label: category, color: color, symbol: symbol);
+        }),
       );
-    }).toList();
+    }
+
+    // Generate legend items for secondary Y-axis (y2Column)
+    final itemsY2 = <LegendItem>[];
+    if (y2Column != null && secondaryGeometries.isNotEmpty) {
+      // Filter categories that have data in the secondary Y column
+      final secondaryCategories = categories.where((category) {
+        return data.any(
+          (row) =>
+              row[colorColumn]?.toString() == category && row[y2Column] != null,
+        );
+      }).toList();
+
+      // Use global index of original categories to maintain consistent colors
+      itemsY2.addAll(
+        secondaryCategories.map((category) {
+          final globalIndex = categories.indexOf(category);
+          final color = colorPalette[globalIndex % colorPalette.length];
+
+          return LegendItem(label: category, color: color, symbol: symbol);
+        }),
+      );
+    }
+
+    return (itemsY, itemsY2);
   }
 
   /// Determine the most appropriate symbol based on chart geometries
   static LegendSymbol _determineSymbolFromGeometries(
-      List<Geometry> geometries) {
+    List<Geometry> geometries,
+  ) {
     if (geometries.isEmpty) return LegendSymbol.circle;
 
     // Priority: check for specific geometry types
