@@ -29,20 +29,21 @@ class _AreaChartExampleState extends State<AreaChartExample> {
   void _generateAreaData() {
     // Generate realistic area chart data - Website Traffic
     areaData = List.generate(12, (i) {
-      final month = [
-        'Jan',
-        'Feb',
-        'Mar',
-        'Apr',
-        'May',
-        'Jun',
-        'Jul',
-        'Aug',
-        'Sep',
-        'Oct',
-        'Nov',
-        'Dec'
-      ][i];
+      final month =
+          [
+            'Jan',
+            'Feb',
+            'Mar',
+            'Apr',
+            'May',
+            'Jun',
+            'Jul',
+            'Aug',
+            'Sep',
+            'Oct',
+            'Nov',
+            'Dec',
+          ][i];
 
       // Simulate seasonal traffic with noise
       final baseTraffic = 5000 + i * 200;
@@ -89,57 +90,61 @@ class _AreaChartExampleState extends State<AreaChartExample> {
               borderRadius: BorderRadius.circular(8),
               border: Border.all(color: widget.theme.borderColor),
             ),
-            child: CristalyseChart()
-                .data(areaData)
-                .mapping(x: 'month', y: 'traffic')
-                .geomArea(
-                  strokeWidth: 2.0,
-                  alpha: 0.3,
-                  fillArea: true,
-                  color: widget.colorPalette.first,
-                )
-                .scaleXOrdinal()
-                .scaleYContinuous(min: 0)
-                .theme(ChartTheme(
-                  backgroundColor: widget.theme.backgroundColor,
-                  primaryColor: widget.theme.primaryColor,
-                  colorPalette: widget.colorPalette,
-                  gridColor: widget.theme.gridColor,
-                  borderColor: widget.theme.borderColor,
-                  axisColor: widget.theme.axisColor,
-                  axisLabelStyle: widget.theme.axisLabelStyle,
-                  axisTextStyle: widget.theme.axisTextStyle,
-                  pointSizeMin: widget.theme.pointSizeMin,
-                  pointSizeMax: widget.theme.pointSizeMax,
-                  pointSizeDefault: widget.theme.pointSizeDefault,
-                  padding: widget.theme.padding,
-                  plotBackgroundColor: widget.theme.plotBackgroundColor,
-                  gridWidth: widget.theme.gridWidth,
-                  axisWidth: widget.theme.axisWidth,
-                ))
-                .animate(
-                  duration: const Duration(milliseconds: 1200),
-                  curve: Curves.easeInOutCubic,
-                )
-                .interaction(
-                  tooltip: TooltipConfig(
-                    builder: (point) => Container(
-                      padding: const EdgeInsets.all(8),
-                      decoration: BoxDecoration(
-                        color: Colors.black.withValues(alpha: 0.8),
-                        borderRadius: BorderRadius.circular(4),
+            child:
+                CristalyseChart()
+                    .data(areaData)
+                    .mapping(x: 'month', y: 'traffic')
+                    .geomArea(
+                      strokeWidth: 2.0,
+                      alpha: 0.3,
+                      fillArea: true,
+                      color: widget.colorPalette.first,
+                    )
+                    .scaleXOrdinal()
+                    .scaleYContinuous(min: 0)
+                    .theme(
+                      ChartTheme(
+                        backgroundColor: widget.theme.backgroundColor,
+                        primaryColor: widget.theme.primaryColor,
+                        colorPalette: widget.colorPalette,
+                        gridColor: widget.theme.gridColor,
+                        borderColor: widget.theme.borderColor,
+                        axisColor: widget.theme.axisColor,
+                        axisLabelStyle: widget.theme.axisLabelStyle,
+                        axisTextStyle: widget.theme.axisTextStyle,
+                        pointSizeMin: widget.theme.pointSizeMin,
+                        pointSizeMax: widget.theme.pointSizeMax,
+                        pointSizeDefault: widget.theme.pointSizeDefault,
+                        padding: widget.theme.padding,
+                        plotBackgroundColor: widget.theme.plotBackgroundColor,
+                        gridWidth: widget.theme.gridWidth,
+                        axisWidth: widget.theme.axisWidth,
                       ),
-                      child: Text(
-                        '${point.getDisplayValue('month')}: ${point.getDisplayValue('traffic')} visitors',
-                        style: const TextStyle(
-                          color: Colors.white,
-                          fontSize: 12,
-                        ),
+                    )
+                    .animate(
+                      duration: const Duration(milliseconds: 1200),
+                      curve: Curves.easeInOutCubic,
+                    )
+                    .interaction(
+                      tooltip: TooltipConfig(
+                        builder:
+                            (point) => Container(
+                              padding: const EdgeInsets.all(8),
+                              decoration: BoxDecoration(
+                                color: Colors.black.withValues(alpha: 0.8),
+                                borderRadius: BorderRadius.circular(4),
+                              ),
+                              child: Text(
+                                '${point.getDisplayValue('month')}: ${point.getDisplayValue('traffic')} visitors',
+                                style: const TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 12,
+                                ),
+                              ),
+                            ),
                       ),
-                    ),
-                  ),
-                )
-                .build(),
+                    )
+                    .build(),
           ),
 
           const SizedBox(height: 24),
@@ -163,71 +168,71 @@ class _AreaChartExampleState extends State<AreaChartExample> {
               borderRadius: BorderRadius.circular(8),
               border: Border.all(color: widget.theme.borderColor),
             ),
-            child: CristalyseChart()
-                .data(_createMultiSeriesData())
-                .mapping(x: 'month', y: 'visitors', color: 'platform')
-                .geomArea(
-                  strokeWidth: 1.5,
-                  alpha: 0.4,
-                  fillArea: true,
-                )
-                .scaleXOrdinal()
-                .scaleYContinuous(min: 0)
-                .theme(ChartTheme(
-                  backgroundColor: widget.theme.backgroundColor,
-                  primaryColor: widget.theme.primaryColor,
-                  colorPalette: widget.colorPalette,
-                  gridColor: widget.theme.gridColor,
-                  borderColor: widget.theme.borderColor,
-                  axisColor: widget.theme.axisColor,
-                  axisLabelStyle: widget.theme.axisLabelStyle,
-                  axisTextStyle: widget.theme.axisTextStyle,
-                  pointSizeMin: widget.theme.pointSizeMin,
-                  pointSizeMax: widget.theme.pointSizeMax,
-                  pointSizeDefault: widget.theme.pointSizeDefault,
-                  padding: widget.theme.padding,
-                  plotBackgroundColor: widget.theme.plotBackgroundColor,
-                  gridWidth: widget.theme.gridWidth,
-                  axisWidth: widget.theme.axisWidth,
-                ))
-                .legend(position: LegendPosition.topRight)
-                .animate(
-                  duration: const Duration(milliseconds: 1400),
-                  curve: Curves.easeInOutCubic,
-                )
-                .interaction(
-                  tooltip: TooltipConfig(
-                    builder: (point) => Container(
-                      padding: const EdgeInsets.all(8),
-                      decoration: BoxDecoration(
-                        color: Colors.black.withValues(alpha: 0.8),
-                        borderRadius: BorderRadius.circular(4),
+            child:
+                CristalyseChart()
+                    .data(_createMultiSeriesData())
+                    .mapping(x: 'month', y: 'visitors', color: 'platform')
+                    .geomArea(strokeWidth: 1.5, alpha: 0.4, fillArea: true)
+                    .scaleXOrdinal()
+                    .scaleYContinuous(min: 0)
+                    .theme(
+                      ChartTheme(
+                        backgroundColor: widget.theme.backgroundColor,
+                        primaryColor: widget.theme.primaryColor,
+                        colorPalette: widget.colorPalette,
+                        gridColor: widget.theme.gridColor,
+                        borderColor: widget.theme.borderColor,
+                        axisColor: widget.theme.axisColor,
+                        axisLabelStyle: widget.theme.axisLabelStyle,
+                        axisTextStyle: widget.theme.axisTextStyle,
+                        pointSizeMin: widget.theme.pointSizeMin,
+                        pointSizeMax: widget.theme.pointSizeMax,
+                        pointSizeDefault: widget.theme.pointSizeDefault,
+                        padding: widget.theme.padding,
+                        plotBackgroundColor: widget.theme.plotBackgroundColor,
+                        gridWidth: widget.theme.gridWidth,
+                        axisWidth: widget.theme.axisWidth,
                       ),
-                      child: Column(
-                        mainAxisSize: MainAxisSize.min,
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Text(
-                            '${point.getDisplayValue('platform')} Traffic',
-                            style: const TextStyle(
-                              color: Colors.white,
-                              fontSize: 12,
-                              fontWeight: FontWeight.bold,
+                    )
+                    .legend(position: LegendPosition.topRight)
+                    .animate(
+                      duration: const Duration(milliseconds: 1400),
+                      curve: Curves.easeInOutCubic,
+                    )
+                    .interaction(
+                      tooltip: TooltipConfig(
+                        builder:
+                            (point) => Container(
+                              padding: const EdgeInsets.all(8),
+                              decoration: BoxDecoration(
+                                color: Colors.black.withValues(alpha: 0.8),
+                                borderRadius: BorderRadius.circular(4),
+                              ),
+                              child: Column(
+                                mainAxisSize: MainAxisSize.min,
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                children: [
+                                  Text(
+                                    '${point.getDisplayValue('platform')} Traffic',
+                                    style: const TextStyle(
+                                      color: Colors.white,
+                                      fontSize: 12,
+                                      fontWeight: FontWeight.bold,
+                                    ),
+                                  ),
+                                  Text(
+                                    '${point.getDisplayValue('month')}: ${point.getDisplayValue('visitors')} visitors',
+                                    style: const TextStyle(
+                                      color: Colors.white,
+                                      fontSize: 12,
+                                    ),
+                                  ),
+                                ],
+                              ),
                             ),
-                          ),
-                          Text(
-                            '${point.getDisplayValue('month')}: ${point.getDisplayValue('visitors')} visitors',
-                            style: const TextStyle(
-                              color: Colors.white,
-                              fontSize: 12,
-                            ),
-                          ),
-                        ],
                       ),
-                    ),
-                  ),
-                )
-                .build(),
+                    )
+                    .build(),
           ),
 
           const SizedBox(height: 24),
@@ -247,67 +252,71 @@ class _AreaChartExampleState extends State<AreaChartExample> {
               borderRadius: BorderRadius.circular(8),
               border: Border.all(color: widget.theme.borderColor),
             ),
-            child: CristalyseChart()
-                .data(areaData)
-                .mapping(x: 'month', y: 'traffic')
-                .geomArea(
-                  strokeWidth: 0, // No stroke for background area
-                  alpha: 0.2,
-                  fillArea: true,
-                  color: widget.colorPalette.first,
-                )
-                .geomLine(
-                  strokeWidth: 3.0,
-                  alpha: 1.0,
-                  color: widget.colorPalette.first,
-                )
-                .geomPoint(
-                  size: 6.0,
-                  alpha: 1.0,
-                  color: widget.colorPalette.first,
-                )
-                .scaleXOrdinal()
-                .scaleYContinuous(min: 0)
-                .theme(ChartTheme(
-                  backgroundColor: widget.theme.backgroundColor,
-                  primaryColor: widget.theme.primaryColor,
-                  colorPalette: widget.colorPalette,
-                  gridColor: widget.theme.gridColor,
-                  borderColor: widget.theme.borderColor,
-                  axisColor: widget.theme.axisColor,
-                  axisLabelStyle: widget.theme.axisLabelStyle,
-                  axisTextStyle: widget.theme.axisTextStyle,
-                  pointSizeMin: widget.theme.pointSizeMin,
-                  pointSizeMax: widget.theme.pointSizeMax,
-                  pointSizeDefault: widget.theme.pointSizeDefault,
-                  padding: widget.theme.padding,
-                  plotBackgroundColor: widget.theme.plotBackgroundColor,
-                  gridWidth: widget.theme.gridWidth,
-                  axisWidth: widget.theme.axisWidth,
-                ))
-                .animate(
-                  duration: const Duration(milliseconds: 1600),
-                  curve: Curves.easeInOutCubic,
-                )
-                .interaction(
-                  tooltip: TooltipConfig(
-                    builder: (point) => Container(
-                      padding: const EdgeInsets.all(8),
-                      decoration: BoxDecoration(
-                        color: Colors.black.withValues(alpha: 0.8),
-                        borderRadius: BorderRadius.circular(4),
+            child:
+                CristalyseChart()
+                    .data(areaData)
+                    .mapping(x: 'month', y: 'traffic')
+                    .geomArea(
+                      strokeWidth: 0, // No stroke for background area
+                      alpha: 0.2,
+                      fillArea: true,
+                      color: widget.colorPalette.first,
+                    )
+                    .geomLine(
+                      strokeWidth: 3.0,
+                      alpha: 1.0,
+                      color: widget.colorPalette.first,
+                    )
+                    .geomPoint(
+                      size: 6.0,
+                      alpha: 1.0,
+                      color: widget.colorPalette.first,
+                    )
+                    .scaleXOrdinal()
+                    .scaleYContinuous(min: 0)
+                    .theme(
+                      ChartTheme(
+                        backgroundColor: widget.theme.backgroundColor,
+                        primaryColor: widget.theme.primaryColor,
+                        colorPalette: widget.colorPalette,
+                        gridColor: widget.theme.gridColor,
+                        borderColor: widget.theme.borderColor,
+                        axisColor: widget.theme.axisColor,
+                        axisLabelStyle: widget.theme.axisLabelStyle,
+                        axisTextStyle: widget.theme.axisTextStyle,
+                        pointSizeMin: widget.theme.pointSizeMin,
+                        pointSizeMax: widget.theme.pointSizeMax,
+                        pointSizeDefault: widget.theme.pointSizeDefault,
+                        padding: widget.theme.padding,
+                        plotBackgroundColor: widget.theme.plotBackgroundColor,
+                        gridWidth: widget.theme.gridWidth,
+                        axisWidth: widget.theme.axisWidth,
                       ),
-                      child: Text(
-                        '${point.getDisplayValue('month')}: ${point.getDisplayValue('traffic')} visitors',
-                        style: const TextStyle(
-                          color: Colors.white,
-                          fontSize: 12,
-                        ),
+                    )
+                    .animate(
+                      duration: const Duration(milliseconds: 1600),
+                      curve: Curves.easeInOutCubic,
+                    )
+                    .interaction(
+                      tooltip: TooltipConfig(
+                        builder:
+                            (point) => Container(
+                              padding: const EdgeInsets.all(8),
+                              decoration: BoxDecoration(
+                                color: Colors.black.withValues(alpha: 0.8),
+                                borderRadius: BorderRadius.circular(4),
+                              ),
+                              child: Text(
+                                '${point.getDisplayValue('month')}: ${point.getDisplayValue('traffic')} visitors',
+                                style: const TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 12,
+                                ),
+                              ),
+                            ),
                       ),
-                    ),
-                  ),
-                )
-                .build(),
+                    )
+                    .build(),
           ),
         ],
       ),
