@@ -27,96 +27,107 @@ Widget buildInteractiveScatterTab(
         const SizedBox(height: 16),
         SizedBox(
           height: 400,
-          child: CristalyseChart()
-              .data(data)
-              .mapping(x: 'x', y: 'y', color: 'category', size: 'size')
-              .geomPoint(
-                alpha: 0.8,
-                size: 4.0 + sliderValue * 8.0,
-              )
-              .scaleXContinuous()
-              .scaleYContinuous()
-              .theme(currentTheme.copyWith(
-                pointSizeMax: 2.0 + sliderValue * 20.0,
-              ))
-              .interaction(
-                tooltip: TooltipConfig(
-                  builder: (point) {
-                    return Container(
-                      padding: const EdgeInsets.all(12),
-                      decoration: BoxDecoration(
-                        color: Colors.black.withValues(alpha: 0.9),
-                        borderRadius: BorderRadius.circular(8),
-                        border: Border.all(color: Colors.white24, width: 1),
-                        boxShadow: const [
-                          BoxShadow(
-                            color: Colors.black26,
-                            blurRadius: 12.0,
-                            offset: Offset(0, 4),
+          child:
+              CristalyseChart()
+                  .data(data)
+                  .mapping(x: 'x', y: 'y', color: 'category', size: 'size')
+                  .geomPoint(alpha: 0.8, size: 4.0 + sliderValue * 8.0)
+                  .scaleXContinuous()
+                  .scaleYContinuous()
+                  .theme(
+                    currentTheme.copyWith(
+                      pointSizeMax: 2.0 + sliderValue * 20.0,
+                    ),
+                  )
+                  .interaction(
+                    tooltip: TooltipConfig(
+                      builder: (point) {
+                        return Container(
+                          padding: const EdgeInsets.all(12),
+                          decoration: BoxDecoration(
+                            color: Colors.black.withValues(alpha: 0.9),
+                            borderRadius: BorderRadius.circular(8),
+                            border: Border.all(color: Colors.white24, width: 1),
+                            boxShadow: const [
+                              BoxShadow(
+                                color: Colors.black26,
+                                blurRadius: 12.0,
+                                offset: Offset(0, 4),
+                              ),
+                            ],
                           ),
-                        ],
-                      ),
-                      child: Column(
-                        mainAxisSize: MainAxisSize.min,
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Text(
-                            'Sales Data',
-                            style: TextStyle(
-                              color: currentTheme.primaryColor,
-                              fontWeight: FontWeight.bold,
-                              fontSize: 14,
-                            ),
+                          child: Column(
+                            mainAxisSize: MainAxisSize.min,
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text(
+                                'Sales Data',
+                                style: TextStyle(
+                                  color: currentTheme.primaryColor,
+                                  fontWeight: FontWeight.bold,
+                                  fontSize: 14,
+                                ),
+                              ),
+                              const SizedBox(height: 6),
+                              Text(
+                                'Week: ${point.getDisplayValue('x')}',
+                                style: const TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 12,
+                                ),
+                              ),
+                              Text(
+                                'Revenue: \$${point.getDisplayValue('y')}k',
+                                style: const TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 12,
+                                ),
+                              ),
+                              Text(
+                                'Segment: ${point.getDisplayValue('category')}',
+                                style: const TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 12,
+                                ),
+                              ),
+                              Text(
+                                'Deal Size: ${point.getDisplayValue('size')}',
+                                style: const TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 12,
+                                ),
+                              ),
+                            ],
                           ),
-                          const SizedBox(height: 6),
-                          Text(
-                            'Week: ${point.getDisplayValue('x')}',
-                            style: const TextStyle(
-                                color: Colors.white, fontSize: 12),
-                          ),
-                          Text(
-                            'Revenue: \$${point.getDisplayValue('y')}k',
-                            style: const TextStyle(
-                                color: Colors.white, fontSize: 12),
-                          ),
-                          Text(
-                            'Segment: ${point.getDisplayValue('category')}',
-                            style: const TextStyle(
-                                color: Colors.white, fontSize: 12),
-                          ),
-                          Text(
-                            'Deal Size: ${point.getDisplayValue('size')}',
-                            style: const TextStyle(
-                                color: Colors.white, fontSize: 12),
-                          ),
-                        ],
-                      ),
-                    );
-                  },
-                  showDelay: const Duration(
-                      milliseconds: 10), // Almost instant for smooth switching
-                  hideDelay:
-                      const Duration(milliseconds: 1500), // Much longer to hide
-                  followPointer: false, // Disable to prevent pan interference
-                ),
-                hover: const HoverConfig(
-                  hitTestRadius: 30.0, // Very generous hit area
-                  debounce: Duration(milliseconds: 50),
-                ),
-                click: ClickConfig(
-                  onTap: (point) {
-                    // In a real app, you'd navigate to details or show a dialog
-                    debugPrint('Tapped on data point: ${point.data}');
-                  },
-                  hitTestRadius: 35.0, // Even more generous for taps
-                ),
-              )
-              .legend(position: LegendPosition.top)
-              .animate(
-                duration: const Duration(milliseconds: 800),
-                curve: Curves.elasticOut,
-              )
-              .build(),
+                        );
+                      },
+                      showDelay: const Duration(
+                        milliseconds: 10,
+                      ), // Almost instant for smooth switching
+                      hideDelay: const Duration(
+                        milliseconds: 1500,
+                      ), // Much longer to hide
+                      followPointer:
+                          false, // Disable to prevent pan interference
+                    ),
+                    hover: const HoverConfig(
+                      hitTestRadius: 30.0, // Very generous hit area
+                      debounce: Duration(milliseconds: 50),
+                    ),
+                    click: ClickConfig(
+                      onTap: (point) {
+                        // In a real app, you'd navigate to details or show a dialog
+                        debugPrint('Tapped on data point: ${point.data}');
+                      },
+                      hitTestRadius: 35.0, // Even more generous for taps
+                    ),
+                  )
+                  .legend(position: LegendPosition.top)
+                  .animate(
+                    duration: const Duration(milliseconds: 800),
+                    curve: Curves.elasticOut,
+                  )
+                  .build(),
         ),
         const SizedBox(height: 16),
         const _InteractionGuide(),
@@ -193,11 +204,7 @@ class _InteractionGuide extends StatelessWidget {
     return Row(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        Icon(
-          icon,
-          size: 16,
-          color: Colors.blue,
-        ),
+        Icon(icon, size: 16, color: Colors.blue),
         const SizedBox(width: 12),
         Expanded(
           child: Column(
@@ -237,11 +244,13 @@ class TooltipExamples {
   /// Multi-column tooltip
   static Widget multi() {
     return CristalyseChart()
-        .tooltip(DefaultTooltips.multi({
-          'revenue': 'Revenue',
-          'deals': 'Deal Count',
-          'conversion': 'Conversion Rate',
-        }))
+        .tooltip(
+          DefaultTooltips.multi({
+            'revenue': 'Revenue',
+            'deals': 'Deal Count',
+            'conversion': 'Conversion Rate',
+          }),
+        )
         .build();
   }
 
@@ -282,22 +291,23 @@ class TooltipExamples {
 
   /// With click actions
   static Widget withActions(BuildContext context) {
-    return CristalyseChart()
-        .tooltip(DefaultTooltips.simple('revenue'))
-        .onClick((point) {
-      showDialog(
-        context: context,
-        builder: (context) => AlertDialog(
-          title: const Text('Data Point Details'),
-          content: Text('Selected: ${point.data}'),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.pop(context),
-              child: const Text('Close'),
-            ),
-          ],
-        ),
-      );
-    }).build();
+    return CristalyseChart().tooltip(DefaultTooltips.simple('revenue')).onClick(
+      (point) {
+        showDialog(
+          context: context,
+          builder:
+              (context) => AlertDialog(
+                title: const Text('Data Point Details'),
+                content: Text('Selected: ${point.data}'),
+                actions: [
+                  TextButton(
+                    onPressed: () => Navigator.pop(context),
+                    child: const Text('Close'),
+                  ),
+                ],
+              ),
+        );
+      },
+    ).build();
   }
 }
