@@ -156,8 +156,11 @@ class CristalyseChart {
   /// - [geomProgress] for configuring progress bar appearance
   /// - [ProgressOrientation] for bar orientation options
   /// - [ProgressStyle] for styling options (filled, gradient, striped, etc.)
-  CristalyseChart mappingProgress(
-      {String? value, String? label, String? category}) {
+  CristalyseChart mappingProgress({
+    String? value,
+    String? label,
+    String? category,
+  }) {
     _progressValueColumn = value;
     _progressLabelColumn = label;
     _progressCategoryColumn = category;
@@ -538,9 +541,14 @@ class CristalyseChart {
     double? max,
     LabelCallback? labels,
     String? title,
+    TickConfig? tickConfig,
   }) {
-    _xScale =
-        LinearScale(limits: (min, max), labelFormatter: labels, title: title);
+    _xScale = LinearScale(
+      limits: (min, max),
+      labelFormatter: labels,
+      title: title,
+      tickConfig: tickConfig,
+    );
     return this;
   }
 
@@ -550,9 +558,14 @@ class CristalyseChart {
     double? max,
     LabelCallback? labels,
     String? title,
+    TickConfig? tickConfig,
   }) {
-    _yScale =
-        LinearScale(limits: (min, max), labelFormatter: labels, title: title);
+    _yScale = LinearScale(
+      limits: (min, max),
+      labelFormatter: labels,
+      title: title,
+      tickConfig: tickConfig,
+    );
     return this;
   }
 
@@ -560,16 +573,22 @@ class CristalyseChart {
   ///
   /// Example:
   /// ```dart
-  /// chart.scaleY2Continuous(min: 0, max: 100, title: 'Conversion Rate (%)') // For percentage data
+  /// chart.scaleY2Continuous(min: 0, max: 100, title: 'Conversion Rate (%)', tickConfig: TickConfig(simpleLinear: true)) // For percentage data
+  /// chart.scaleY2Continuous(min: 0, max: 100, title: 'Conversion Rate (%)', tickConfig: TickConfig(ticks: [0, 25, 50, 75, 100])) // For percentage data
   /// ```
   CristalyseChart scaleY2Continuous({
     double? min,
     double? max,
     LabelCallback? labels,
     String? title,
+    TickConfig? tickConfig,
   }) {
-    _y2Scale =
-        LinearScale(limits: (min, max), labelFormatter: labels, title: title);
+    _y2Scale = LinearScale(
+      limits: (min, max),
+      labelFormatter: labels,
+      title: title,
+      tickConfig: tickConfig,
+    );
     return this;
   }
 
@@ -653,7 +672,8 @@ class CristalyseChart {
     if ((categoryColors == null || categoryColors.isEmpty) &&
         (categoryGradients == null || categoryGradients.isEmpty)) {
       throw ArgumentError(
-          'Either categoryColors or categoryGradients must be provided and non-empty');
+        'Either categoryColors or categoryGradients must be provided and non-empty',
+      );
     }
     if (_colorColumn != null) {
       final String colorColumn = _colorColumn ?? '';
@@ -661,20 +681,26 @@ class CristalyseChart {
       // Apply solid colors if provided
       if (categoryColors != null && categoryColors.isNotEmpty) {
         _theme = _theme.customPalette(
-            data: _data, color: colorColumn, categoryColors: categoryColors);
+          data: _data,
+          color: colorColumn,
+          categoryColors: categoryColors,
+        );
       }
 
       // Apply gradients if provided
       if (categoryGradients != null && categoryGradients.isNotEmpty) {
         _theme = _theme.customGradientPalette(
-            data: _data,
-            color: colorColumn,
-            categoryGradients: categoryGradients);
+          data: _data,
+          color: colorColumn,
+          categoryGradients: categoryGradients,
+        );
       }
     } else {
-      throw ArgumentError("'color' argument is missing from .mapping. \n"
-          "The correct code should look like this .mapping(x:'', y='', color='')\n"
-          "If you don't wish to add a category column, remove customPalette() from your CristalyseChart declaration code.\n\n");
+      throw ArgumentError(
+        "'color' argument is missing from .mapping. \n"
+        "The correct code should look like this .mapping(x:'', y='', color='')\n"
+        "If you don't wish to add a category column, remove customPalette() from your CristalyseChart declaration code.\n\n",
+      );
     }
     return this;
   }
@@ -746,6 +772,7 @@ class CristalyseChart {
     HoverConfig? hover,
     ClickConfig? click,
     PanConfig? pan,
+    ZoomConfig? zoom,
     bool enabled = true,
   }) {
     _interaction = ChartInteraction(
@@ -753,6 +780,7 @@ class CristalyseChart {
       hover: hover,
       click: click,
       pan: pan,
+      zoom: zoom,
       enabled: enabled,
     );
     return this;
@@ -810,6 +838,22 @@ class CristalyseChart {
     return this;
   }
 
+  /// Quick zoom setup (defaults to X-axis zooming)
+  ///
+  /// Example:
+  /// ```dart
+  /// chart.onZoom((info) {
+  ///   print('Zoom scale: ${info.scaleX}');
+  /// });
+  /// ```
+  CristalyseChart onZoom(ZoomCallback callback, {ZoomAxis axis = ZoomAxis.x}) {
+    _interaction = ChartInteraction(
+      zoom: ZoomConfig(enabled: true, axes: axis, onZoomUpdate: callback),
+      enabled: true,
+    );
+    return this;
+  }
+
   /// Quick click setup
   ///
   /// Example:
@@ -828,6 +872,7 @@ class CristalyseChart {
   ///
   /// Automatically generates a legend based on the color mapping column.
   /// Only works when a `color` mapping is defined in `.mapping()`.
+  /// Show titles for y and y2 axes titles for the legend group if showTitles is true.
   ///
   /// Basic usage:
   /// ```dart
@@ -897,6 +942,7 @@ class CristalyseChart {
     bool? interactive,
     Set<String>? hiddenCategories,
     void Function(String category, bool visible)? onToggle,
+    bool? showTitles,
   }) {
     _legendConfig = LegendConfig(
       position: position ?? LegendPosition.topRight,
@@ -913,6 +959,7 @@ class CristalyseChart {
       interactive: interactive ?? false,
       hiddenCategories: hiddenCategories,
       onToggle: onToggle,
+      showTitles: showTitles ?? false,
     );
     return this;
   }
@@ -1116,21 +1163,22 @@ extension ChartThemeExtension on ChartTheme {
   /// ```
   ///
   /// Returns a new [ChartTheme] with the custom color palette applied.
-  ChartTheme customPalette(
-      {required List<Map<String, dynamic>> data,
-      required String color,
-      required Map<String, Color> categoryColors}) {
+  ChartTheme customPalette({
+    required List<Map<String, dynamic>> data,
+    required String color,
+    required Map<String, Color> categoryColors,
+  }) {
     // Extract unique categories
     final categories = data.map((d) => d[color] as String).toSet().toList();
     // Build color palette in the order categories appear
     final colorPalette = categories
-        .map((category) =>
-            categoryColors[category] ??
-            this.colorPalette[categories.indexOf(category)])
+        .map(
+          (category) =>
+              categoryColors[category] ??
+              this.colorPalette[categories.indexOf(category)],
+        )
         .toList();
-    return copyWith(
-      colorPalette: colorPalette,
-    );
+    return copyWith(colorPalette: colorPalette);
   }
 
   /// Creates a new [ChartTheme] with custom gradients for specific categories
@@ -1180,8 +1228,6 @@ extension ChartThemeExtension on ChartTheme {
     required String color,
     required Map<String, Gradient> categoryGradients,
   }) {
-    return copyWith(
-      categoryGradients: categoryGradients,
-    );
+    return copyWith(categoryGradients: categoryGradients);
   }
 }
